feat: Add all logos + Main logo refactoring + Sort logos refactoring
