+++ packages/client/src/assets/images/index.ts
@@ -1,0 +1,63 @@
+import angular from './logos/angular.svg'
+import bootstrap from './logos/bootstrap.svg'
+import css from './logos/css.svg'
+import docker from './logos/docker.svg'
+import eslint from './logos/eslint.svg'
+import figma from './logos/figma.svg'
+import git from './logos/git.svg'
+import github from './logos/github.svg'
+import gitlab from './logos/gitlab.svg'
+import go from './logos/go.svg'
+import graphql from './logos/graphql.svg'
+import html from './logos/html.svg'
+import jest from './logos/jest.svg'
+import js from './logos/js.svg'
+import laravel from './logos/laravel.svg'
+import mui from './logos/mui.svg'
+import node from './logos/node.svg'
+import npm from './logos/npm.svg'
+import nuxt from './logos/nuxt.svg'
+import python from './logos/python.svg'
+import prettier from './logos/prettier.svg'
+import pug from './logos/pug.svg'
+import react from './logos/react.svg'
+import redux from './logos/redux.svg'
+import sass from './logos/sass.svg'
+import ts from './logos/ts.svg'
+import vite from './logos/vite.svg'
+import vue from './logos/vue.svg'
+import webpack from './logos/webpack.svg'
+import yarn from './logos/yarn.svg'
+
+export {
+  angular,
+  bootstrap,
+  css,
+  docker,
+  eslint,
+  figma,
+  git,
+  github,
+  gitlab,
+  go,
+  graphql,
+  html,
+  jest,
+  js,
+  laravel,
+  mui,
+  node,
+  npm,
+  nuxt,
+  python,
+  prettier,
+  pug,
+  react,
+  redux,
+  sass,
+  ts,
+  vite,
+  vue,
+  webpack,
+  yarn,
+}
