--- packages/client/src/assets/images/logos/index.ts
@@ -1,10 +1,0 @@
-import angular from './angular.svg'
-import github from './github.svg'
-import gitlab from './gitlab.svg'
-import js from './js.svg'
-import kraken from './kraken.svg'
-import lighthouse from './lighthouse.svg'
-import react from './react.svg'
-import sass from './sass.svg'
-
-export { angular, github, gitlab, js, kraken, lighthouse, react, sass }
