+++ packages/client/src/assets/images/other/hooray.gif
