+++ packages/client/src/components/Avatar/Avatar.module.scss
@@ -19,7 +19,7 @@
   display: inline-block;
   width: 44px;
   height: 44px;
-  background: url("@/assets/images/image-icon.svg") center/50% var(--color-background-secondary) no-repeat;
+  background: url("@/assets/images/other/image-icon.svg") center/50% var(--color-background-secondary) no-repeat;
   border-radius: 50%;
   cursor: pointer;
   margin: 0 auto;
