+++ packages/client/src/config/gameConfig.ts
@@ -1,23 +1,67 @@
 import {
   angular,
+  bootstrap,
+  css,
+  docker,
+  eslint,
+  figma,
+  git,
   github,
   gitlab,
+  go,
+  graphql,
+  html,
+  jest,
   js,
-  kraken,
-  lighthouse,
+  laravel,
+  mui,
+  node,
+  npm,
+  nuxt,
+  python,
+  prettier,
+  pug,
   react,
+  redux,
   sass,
-} from '@/assets/images/logos/index'
+  ts,
+  vite,
+  vue,
+  webpack,
+  yarn,
+} from '@/assets/images'
 
 export const allIcons = [
   angular,
+  bootstrap,
+  css,
+  docker,
+  eslint,
+  figma,
+  git,
   github,
   gitlab,
+  go,
+  graphql,
+  html,
+  jest,
   js,
-  kraken,
-  lighthouse,
+  laravel,
+  mui,
+  node,
+  npm,
+  nuxt,
+  python,
+  prettier,
+  pug,
   react,
+  redux,
   sass,
+  ts,
+  vite,
+  vue,
+  webpack,
+  yarn,
 ]
 
 const getCardSize = (cols: number) => (cols === 4 ? 120 : 85)
@@ -37,6 +81,8 @@ const iconsCount: Record<CardsCount, number> = {
   [CardsCount.M]: 18,
   [CardsCount.L]: 30,
 }
+
+export const randomSortedIcons = allIcons.sort(() => Math.random() - 0.5)
 
 export const getGameConfig = (gameCols: number) => {
   const rows = getRowsSize(gameCols)
@@ -59,13 +105,16 @@ export const getGameConfig = (gameCols: number) => {
   // Получаем нужное колличество иконок в зависимости от gameConfig.cols * gameConfig.rows
   const computedIconsCount =
     computedTotalGameCards === 60
-      ? allIcons
-      : allIcons.slice(0, iconsCount[computedTotalGameCards as CardsCount])
+      ? randomSortedIcons
+      : randomSortedIcons.slice(
+          0,
+          iconsCount[computedTotalGameCards as CardsCount]
+        )
 
   return {
     cols: gameCols,
     gameConfig: computedGameConfig,
-    getIconsCount: computedIconsCount,
+    icons: computedIconsCount,
     rows,
     totalGameCards: computedTotalGameCards,
     FPS,
